+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Released]
 
+## [1.1.0] - 2022-02-21
+
+### Added
+- Schedule log cleaner cronjob on plugin activation. The cronjob deletes rows from log table older than the threshold set.
+
+### Fixed
+- Fixed autoloading, added composer test + more #17
+
 ## [1.0.1] - 2022-02-03
 
 ### Fixed